Move registering field indexes to the noderefutil package
